+++ TidalPy/structures_x/layers/__init__.py
@@ -3,5 +3,6 @@
 from TidalPy.structures_x.layers.base import BaseLayer
 from TidalPy.structures_x.layers.physics import PhysicsLayer
 from TidalPy.structures_x.layers.solidliquid import SolidLiquidLayer
+from TidalPy.structures_x.layers.gas import GasLayer
 
-__all__ = ["BaseLayer", "PhysicsLayer", "SolidLiquidLayer"]
+__all__ = ["BaseLayer", "PhysicsLayer", "SolidLiquidLayer", "GasLayer"]